+++ src/EventManager/EventManager.ts
@@ -1,5 +1,6 @@
 import invariant from 'tiny-invariant';
 import { AnyizeType } from '../AnyizeCSM';
+import { eventListenerFactory } from './eventListener'
 import { IEventPageTarget, IEventStepTarget, IStepTargetModifiers, IPageTargetModifiers, IOrderStatusModifiers } from '../EventTarget/types';
 import { debug } from '../shared/logger';
 import { 
@@ -17,9 +18,18 @@ let prevStep: typeof window.Shopify.Checkout["step"];
  */
 export class EventManager {
     private _debug: boolean;
+    private _addEventListener: ReturnType<typeof eventListenerFactory>["addEventListener"];
+    private _removeEventListener: ReturnType<typeof eventListenerFactory>["removeEventListener"];
 
     public constructor(options: IOptions) {
         this._debug = options.debug;
+        const {
+            addEventListener,
+            removeEventListener
+        } = eventListenerFactory(options.jQuery)
+
+        this._addEventListener = addEventListener;
+        this._removeEventListener = removeEventListener;
     }
 
     private _trigger = (callbacks: Function[], eventName: string) => {
@@ -50,15 +60,15 @@ export class EventManager {
     }
 
     private _observeAnyRepaint = (callbacks: Function[]): Function => {
-        const eventName = 'Any Repaint';
+        const eventName = `(${PageEvent.PAGE_CHANGE}) Any Repaint`;
         const handleAnyRepaint = () => {
             this._trigger(callbacks, eventName);
         }
 
-        window.addEventListener(PageEvent.PAGE_CHANGE, handleAnyRepaint);
+        this._addEventListener(PageEvent.PAGE_CHANGE, handleAnyRepaint);
 
         return () => {
-            window.removeEventListener(PageEvent.PAGE_CHANGE, handleAnyRepaint);
+            this._removeEventListener(PageEvent.PAGE_CHANGE, handleAnyRepaint);
         }
     }
 
@@ -68,7 +78,7 @@ export class EventManager {
         /**
          * Only trigger callbacks on supported checkout page values
          */
-        const handleAnyPageChange = () => {
+        const handleAnyPageFactory = (eventType: string) => () => {
             const page = window.Shopify.Checkout.page?.toLowerCase();
 
             /**
@@ -86,24 +96,26 @@ export class EventManager {
                 )
             ) {
                 prevPage = page;
-                this._trigger(callbacks, `${eventName} for "${page}"`);
+                this._trigger(callbacks, `(${eventType}) ${eventName} for "${page}"`);
             }
         }
 
-        window.addEventListener(PageEvent.PAGE_LOAD_AND_CHANGE, handleAnyPageChange);
+        const handleAnyPageLoad = handleAnyPageFactory(PageEvent.PAGE_LOAD);
+        const handleAnyPageChange = handleAnyPageFactory(PageEvent.PAGE_CHANGE);
+
+        this._addEventListener(PageEvent.PAGE_LOAD, handleAnyPageLoad);
+        this._addEventListener(PageEvent.PAGE_CHANGE, handleAnyPageChange);
 
         return () => {
-            window.removeEventListener(PageEvent.PAGE_LOAD_AND_CHANGE, handleAnyPageChange);
+            this._removeEventListener(PageEvent.PAGE_LOAD, handleAnyPageLoad);
+            this._removeEventListener(PageEvent.PAGE_CHANGE, handleAnyPageChange);
         }
     }
 
     private _observeAnyStepChange = (callbacks: Function[]): Function => {
         const eventName = 'Any Step Change';
 
-        /**
-         * Only trigger callbacks on supported checkout step values
-         */
-        const handleAnyStepChange = () => {
+        const handleAnyStepFactory = (eventType: string) => () => {
             const step = window.Shopify.Checkout.step?.toLowerCase();
 
             /**
@@ -121,24 +133,29 @@ export class EventManager {
                 )
             ) {
                 prevStep = step;
-                this._trigger(callbacks, `${eventName} for "${step}"`);
+                this._trigger(callbacks, `(${eventType}) ${eventName} for "${step}"`);
             }
         }
 
-        window.addEventListener(PageEvent.PAGE_LOAD_AND_CHANGE, handleAnyStepChange);
+        const handleAnyStepLoad = handleAnyStepFactory(PageEvent.PAGE_LOAD);
+        const handleAnyStepChange = handleAnyStepFactory(PageEvent.PAGE_CHANGE);
+
+        this._addEventListener(PageEvent.PAGE_LOAD, handleAnyStepLoad);
+        this._addEventListener(PageEvent.PAGE_CHANGE, handleAnyStepChange);
 
         return () => {
-            window.removeEventListener(PageEvent.PAGE_LOAD_AND_CHANGE, handleAnyStepChange);
+            this._removeEventListener(PageEvent.PAGE_LOAD, handleAnyStepLoad);
+            this._removeEventListener(PageEvent.PAGE_CHANGE, handleAnyStepChange);
         }
     }
 
     private _observeSpecificStep = (forStep: CheckoutStep, callbacks: Function[], modifiers?: IStepTargetModifiers): Function => {
-        const eventName = `Only for ${forStep} step`;
+        const eventName = `Only for "${forStep}" step`;
         let removeWithAnyPageListener: Function | undefined;
         let removeWithAnyRepaintListener: Function | undefined;
         let removeWithSpecificPageListener: Function | undefined;
         
-        const handleSpecificStep = () => {
+        const handleSpecificStepFactory = (eventType: string) => () => {
             const step = window.Shopify.Checkout.step?.toLowerCase();
 
             if (step === forStep) {
@@ -160,7 +177,7 @@ export class EventManager {
                 }
 
                 // always trigger the callbacks for the first time
-                this._trigger(callbacks, eventName);
+                this._trigger(callbacks, `(${eventType}) ${eventName}`);
             }
             else {
                 removeWithAnyPageListener?.();
@@ -169,20 +186,25 @@ export class EventManager {
             }
         }
 
-        window.addEventListener(PageEvent.PAGE_LOAD_AND_CHANGE, handleSpecificStep);
+        const handleSpecificStepLoad = handleSpecificStepFactory(PageEvent.PAGE_LOAD);
+        const handleSpecificStepChange = handleSpecificStepFactory(PageEvent.PAGE_CHANGE);
+
+        this._addEventListener(PageEvent.PAGE_LOAD, handleSpecificStepLoad);
+        this._addEventListener(PageEvent.PAGE_CHANGE, handleSpecificStepChange);
 
         return () => {
-            window.removeEventListener(PageEvent.PAGE_LOAD_AND_CHANGE, handleSpecificStep);
+            this._removeEventListener(PageEvent.PAGE_LOAD, handleSpecificStepLoad);
+            this._removeEventListener(PageEvent.PAGE_CHANGE, handleSpecificStepChange);
         }
     }
 
     private _observeSpecificPage = (forPage: CheckoutPage, callbacks: Function[], modifiers?: IPageTargetModifiers): Function => {
-        const eventName = `Only for ${forPage} page`;
+        const eventName = `Only for "${forPage}" page`;
         let removeWithAnyStepListener: Function | undefined;
         let removeWithAnyRepaintListener: Function | undefined;
         let removeWithSpecificStepListener: Function | undefined;
 
-        const handleSpecificPage = () => {
+        const handleSpecificPageFactory = (eventType: string) => () => {
             const page = window.Shopify.Checkout.page?.toLowerCase();
 
             if (page === forPage) {
@@ -204,7 +226,7 @@ export class EventManager {
                 }
                     
                 // always trigger the callbacks for the first time
-                this._trigger(callbacks, eventName);
+                this._trigger(callbacks, `(${eventType}) ${eventName}`);
             }
             else {
                 removeWithAnyStepListener?.();
@@ -213,10 +235,15 @@ export class EventManager {
             }
         }  
 
-        window.addEventListener(PageEvent.PAGE_LOAD_AND_CHANGE, handleSpecificPage);
-
+        const handleSpecificPageChange = handleSpecificPageFactory(PageEvent.PAGE_CHANGE);
+        const handleSpecificPageLoad = handleSpecificPageFactory(PageEvent.PAGE_LOAD);
+        
+        this._addEventListener(PageEvent.PAGE_LOAD, handleSpecificPageLoad);
+        this._addEventListener(PageEvent.PAGE_CHANGE, handleSpecificPageChange);
+        
         return () => {
-            window.removeEventListener(PageEvent.PAGE_LOAD_AND_CHANGE, handleSpecificPage);
+            this._removeEventListener(PageEvent.PAGE_LOAD, handleSpecificPageLoad);
+            this._removeEventListener(PageEvent.PAGE_CHANGE, handleSpecificPageChange);
         }
     }
 
@@ -266,15 +293,15 @@ export class EventManager {
                 ) {
                     // trigger on both versions of the Order Status page
                     // as a way without causing any unexpected effects
-                    this._trigger(callbacks, eventName);
+                    this._trigger(callbacks, `(${PageEvent.PAGE_LOAD}) ${eventName}`);
                 }
             }
         }  
 
-        window.addEventListener(PageEvent.PAGE_LOAD, handleOrderStatus);
+        this._addEventListener(PageEvent.PAGE_LOAD, handleOrderStatus);
 
         return () => {
-            window.removeEventListener(PageEvent.PAGE_LOAD, handleOrderStatus);
+            this._removeEventListener(PageEvent.PAGE_LOAD, handleOrderStatus);
         }
     }
-}
+};