+++ src/shared/defaultOptions.ts
@@ -1,5 +1,7 @@
 import { IOptions } from '../shared/types';
+import $ from 'jquery';
 
 export const defaultOptions: IOptions = {
     debug: false,
+    jQuery: $,
 };