+++ src/shared/types.ts
@@ -3,7 +3,8 @@ export interface IOptions {
      * Used to log info during runtime which can be helpful during
      * development or debugging. But should be turned off in production.
      */
-    debug: boolean;
+    debug?: boolean;
+    jQuery?: JQueryStatic;
 }
 
 export interface IBaseCSM {