+++ CHANGELOG.md
@@ -2,6 +2,43 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.0.0](https://github.com/Weffe/shopify-checkout-step-manager/compare/v1.5.0...v2.0.0) (2019-11-11)
+
+
+### Bug Fixes
+
+* fix issue with config options for jQuery ([2635946](https://github.com/Weffe/shopify-checkout-step-manager/commit/2635946ee1ada78dc5d749ca7123eee7831dc042))
+
+## [1.5.0](https://github.com/Weffe/shopify-checkout-step-manager/compare/v1.4.0...v1.5.0) (2019-11-08)
+
+
+### Features
+
+* add better debug information ([532cfec](https://github.com/Weffe/shopify-checkout-step-manager/commit/532cfecda84d3fd9848e6557a9aa0a1a6ea73e2c))
+
+## [1.4.0](https://github.com/Weffe/shopify-checkout-step-manager/compare/v1.3.0...v1.4.0) (2019-11-08)
+
+
+### Features
+
+* allow passing in a custom jQuery instance in options ([8c5a660](https://github.com/Weffe/shopify-checkout-step-manager/commit/8c5a660bb11c0e9d2fcbf672d50974e54e94ecaf))
+
+## [1.3.0](https://github.com/Weffe/shopify-checkout-step-manager/compare/v1.2.0...v1.3.0) (2019-11-08)
+
+
+### Bug Fixes
+
+* fix issue with add/remove event listener to the wrong event name ([8fc1ab1](https://github.com/Weffe/shopify-checkout-step-manager/commit/8fc1ab14c9e07effdbbfa732f3964b27a4d1bf29))
+
+## [1.2.0](https://github.com/Weffe/shopify-checkout-step-manager/compare/v1.1.0...v1.2.0) (2019-11-07)
+
+## [1.1.0](https://github.com/Weffe/shopify-checkout-step-manager/compare/v1.0.0...v1.1.0) (2019-11-07)
+
+
+### Bug Fixes
+
+* fix issue with alias import ([c9d9d7b](https://github.com/Weffe/shopify-checkout-step-manager/commit/c9d9d7b839d9c8d0153d87015a025fea1ee245ee))
+
 ## [1.0.0](https://github.com/Weffe/shopify-checkout-step-manager/compare/v0.1.2...v1.0.0) (2019-11-07)
 
 ### [0.1.2](https://github.com/Weffe/shopify-checkout-step-manager/compare/v0.1.1...v0.1.2) (2019-10-17)